Fixes authentication tests (#381)
* remove Dummy

* cleans up auth tests

* fmt
