+++ .gitignore
@@ -33,3 +33,4 @@ yarn-error.log*
 target
 neardev
 data.json
+/test-results
