--- apps/new/widget/Dummy.jsx
@@ -1,2 +1,0 @@
-// dummy file for pages that are yet to be created
-return <div>dummy</div>;
