+++ package.json
@@ -18,6 +18,7 @@
     "start": "yarn dev:gateway",
     "serve": "webpack serve",
     "test": "npx playwright test",
+    "test:ui": "npx playwright test --ui",
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "prepare": "simple-git-hooks"
   },
