+++ playwright-tests/util/constants.js
@@ -1,3 +1,5 @@
 /**
  * constants
  */
+
+export const ROOT_SRC = "builddao.testnet/widget/Index";
