+++ playwright.config.js
@@ -96,7 +96,7 @@ export default defineConfig({
 
   /* Run your local dev server before starting the tests */
   webServer: {
-    command: "npm run dev",
+    command: "npm run dev:testnet",
     port: 8080,
     reuseExistingServer: !process.env.CI,
   },
